+++ resources-ktx/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 
 - **Context**: added type checks to extensions `resolveColor`
+- **Context**: new extensions `resolveDimension`, `resolveDimensionPixelSize` and `resolveDimensionPixelOffset`
 
 ### Dependencies